Fixing code smells
