--- src/main/java/dev/vitorvidal/characterCreator/CharacterCreatorApplication.java
@@ -1,17 +1,0 @@
-package dev.vitorvidal.characterCreator;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import springfox.documentation.oas.annotations.EnableOpenApi;
-
-@SpringBootApplication
-@EnableWebSecurity
-@EnableOpenApi
-public class CharacterCreatorApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(CharacterCreatorApplication.class, args);
-    }
-
-}
