--- src/main/java/dev/vitorvidal/characterCreator/application/vo/AbilityScoreVO.java
@@ -1,19 +1,0 @@
-package dev.vitorvidal.characterCreator.application.vo;
-
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.util.UUID;
-
-@Data
-@NoArgsConstructor
-public class AbilityScoreVO {
-
-    private UUID id;
-    private int strength;
-    private int dexterity;
-    private int constitution;
-    private int intelligence;
-    private int wisdom;
-    private int charisma;
-}
