--- src/main/java/dev/vitorvidal/characterCreator/domain/model/Class.java
@@ -1,16 +1,0 @@
-package dev.vitorvidal.characterCreator.domain.model;
-
-public enum Class {
-    BARBARIAN,
-    BARD,
-    CLERIC,
-    DRUID,
-    FIGHTER,
-    MONK,
-    PALADIN,
-    RANGER,
-    ROGUE,
-    SORCERER,
-    WARLOCK,
-    WIZARD,
-}
