--- src/main/java/dev/vitorvidal/characterCreator/domain/model/Race.java
@@ -1,15 +1,0 @@
-package dev.vitorvidal.characterCreator.domain.model;
-
-public enum Race {
-    AASIMAR,
-    DRAGONBORN,
-    DWARF,
-    ELF,
-    HALFLING,
-    HUMAN,
-    GENASI,
-    GOLIATH,
-    GNOME,
-    TABAXI,
-    TIEFLING
-}
