--- src/main/java/dev/vitorvidal/characterCreator/domain/repository/CharacterRepository.java
@@ -1,11 +1,0 @@
-package dev.vitorvidal.characterCreator.domain.repository;
-
-import dev.vitorvidal.characterCreator.domain.model.Character;
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.UUID;
-
-@Repository
-public interface CharacterRepository extends CrudRepository<Character, UUID> {
-}
