+++ src/main/java/dev/vitorvidal/charactercreator/application/vo/AbilityScoreVO.java
@@ -1,0 +1,19 @@
+package dev.vitorvidal.charactercreator.application.vo;
+
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.util.UUID;
+
+@Data
+@NoArgsConstructor
+public class AbilityScoreVO {
+
+    private UUID id;
+    private int strength;
+    private int dexterity;
+    private int constitution;
+    private int intelligence;
+    private int wisdom;
+    private int charisma;
+}
