+++ src/main/java/dev/vitorvidal/charactercreator/domain/model/Class.java
@@ -1,0 +1,16 @@
+package dev.vitorvidal.charactercreator.domain.model;
+
+public enum Class {
+    BARBARIAN,
+    BARD,
+    CLERIC,
+    DRUID,
+    FIGHTER,
+    MONK,
+    PALADIN,
+    RANGER,
+    ROGUE,
+    SORCERER,
+    WARLOCK,
+    WIZARD,
+}
