+++ src/main/java/dev/vitorvidal/charactercreator/domain/model/Race.java
@@ -1,0 +1,15 @@
+package dev.vitorvidal.charactercreator.domain.model;
+
+public enum Race {
+    AASIMAR,
+    DRAGONBORN,
+    DWARF,
+    ELF,
+    HALFLING,
+    HUMAN,
+    GENASI,
+    GOLIATH,
+    GNOME,
+    TABAXI,
+    TIEFLING
+}
