+++ src/main/java/dev/vitorvidal/charactercreator/domain/repository/AbilityScoreRepository.java
@@ -1,0 +1,11 @@
+package dev.vitorvidal.charactercreator.domain.repository;
+
+import dev.vitorvidal.charactercreator.domain.model.AbilityScore;
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.UUID;
+
+@Repository
+public interface AbilityScoreRepository extends CrudRepository<AbilityScore, UUID> {
+}
