--- src/test/java/dev/vitorvidal/characterCreator/CharacterCreatorApplicationTests.java
@@ -1,13 +1,0 @@
-package dev.vitorvidal.characterCreator;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class CharacterCreatorApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
