--- src/test/java/dev/vitorvidal/characterCreator/application/service/AbilityScoreServiceTest.java
@@ -1,32 +1,0 @@
-package dev.vitorvidal.characterCreator.application.service;
-
-import org.junit.jupiter.api.Disabled;
-import org.junit.jupiter.api.Test;
-
-class AbilityScoreServiceTest {
-
-    @Test
-    @Disabled("Not Implemented")
-    void getAll() {
-    }
-
-    @Test
-    @Disabled("Not Implemented")
-    void setAbilityScore() {
-    }
-
-    @Test
-    @Disabled("Not Implemented")
-    void getAbilityScoreById() {
-    }
-
-    @Test
-    @Disabled("Not Implemented")
-    void deleteAbilityScore() {
-    }
-
-    @Test
-    @Disabled("Not Implemented")
-    void levelUp() {
-    }
-}
