--- src/test/java/dev/vitorvidal/characterCreator/application/service/CharacterServiceTest.java
@@ -1,32 +1,0 @@
-package dev.vitorvidal.characterCreator.application.service;
-
-import org.junit.jupiter.api.Disabled;
-import org.junit.jupiter.api.Test;
-
-class CharacterServiceTest {
-
-    @Test
-    @Disabled("Not Implemented")
-    void getAll() {
-    }
-
-    @Test
-    @Disabled("Not Implemented")
-    void createCharacter() {
-    }
-
-    @Test
-    @Disabled("Not Implemented")
-    void getCharacterById() {
-    }
-
-    @Test
-    @Disabled("Not Implemented")
-    void updateCharacter() {
-    }
-
-    @Test
-    @Disabled("Not Implemented")
-    void deleteCharacter() {
-    }
-}
