+++ src/test/java/dev/vitorvidal/charactercreator/CharacterCreatorApplicationTests.java
@@ -1,0 +1,13 @@
+package dev.vitorvidal.charactercreator;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class CharacterCreatorApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
